+++ packages/forms/src/widgets/index.js
@@ -1,7 +1,0 @@
-import fields from './fields';
-import templates from './templates';
-
-export default {
-	fields,
-	templates,
-};