+++ tools/scripts-config-jest/test-setup.js
@@ -12,35 +12,14 @@ require('core-js/stable');
 require('regenerator-runtime/runtime');
 require('raf/polyfill');
 
-const warnMessageOptionalDep = (mainDepToMock, depList = []) => {
-	if (depList.length === 0) {
-		console.warn(
-			`JEST MOCK WARN: ${mainDepToMock} is not resolved.` +
-				'\nThis is an optional dependency.' +
-				'\nPlease add it in your dependencies if you need it',
-		);
-	} else if (depList.length > 0) {
-		console.warn(
-			`JEST MOCK WARN: one or more of those deps are not resolved: ${depList.join(', ')}` +
-				'These are optional dependencies but work together.' +
-				`\nIt's needed to mock ${mainDepToMock}` +
-				'\nPlease add them in your dependencies if you need them',
-		);
-	}
-};
-
 try {
 	const jestAxe = require('jest-axe');
 	expect.extend(jestAxe.toHaveNoViolations);
-} catch (e) {
-	warnMessageOptionalDep('jest-axe');
-}
+} catch (e) {}
 
 try {
 	jest.mock('ally.js');
-} catch (e) {
-	warnMessageOptionalDep('ally.js');
-}
+} catch (e) {}
 
 // add missing ResizeObserver
 class ResizeObserver {
@@ -69,6 +48,10 @@ if (!global.self.TextEncoder) {
 
 if (!global.URL?.revokeObjectURL) {
 	global.URL.revokeObjectURL = jest.fn();
+}
+
+if (!global._) {
+	global._ = require('lodash');
 }
 
 // Mock fetch
@@ -147,9 +130,7 @@ try {
 		i18nextMock.addResources = () => {};
 		return i18nextMock;
 	});
-} catch (e) {
-	warnMessageOptionalDep('i18next');
-}
+} catch (e) {}
 
 try {
 	jest.mock('react-i18next', () => {
@@ -201,9 +182,7 @@ try {
 				Array.isArray(children) ? renderNodes(children) : renderNodes([children]),
 		};
 	});
-} catch (e) {
-	warnMessageOptionalDep('react-i18next', ['react-i18next', 'i18next', 'react']);
-}
+} catch (e) {}
 
 try {
 	jest.mock('@talend/design-system', () => {
@@ -260,14 +239,7 @@ try {
 
 		return mocks;
 	});
-} catch {
-	warnMessageOptionalDep('@talend/design-system', [
-		'@talend/design-system',
-		'react',
-		'prop-types',
-		'classnames',
-	]);
-}
+} catch {}
 
 // @floating-ui/react
 // https://github.com/floating-ui/floating-ui/issues/1908