+++ tools/babel-plugin-assets-api/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @talend/babel-plugin-assets-api
 
+## 1.4.0
+
+### Minor Changes
+
+- c3750a1: chore: upgrade dependencies
+
+  Major upgrade for all packages that have a peerDependency on react or react-dom. Those packages now ask for react@18 and react-dom@18.
+
+### Patch Changes
+
+- Updated dependencies [c3750a1]
+  - @talend/module-to-cdn@9.11.0
+
+## 1.3.1
+
+### Patch Changes
+
+- 1abc22f: chore: upgrade dependencies
+- Updated dependencies [1abc22f]
+  - @talend/module-to-cdn@9.10.1
+
 ## 1.3.0
 
 ### Minor Changes