+++ tools/scripts-config-babel/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @talend/scripts-config-babel
 
+## 13.5.0
+
+### Minor Changes
+
+- c3750a1: chore: upgrade dependencies
+
+  Major upgrade for all packages that have a peerDependency on react or react-dom. Those packages now ask for react@18 and react-dom@18.
+
+### Patch Changes
+
+- Updated dependencies [c3750a1]
+  - @talend/babel-plugin-import-from-index@1.8.0
+  - @talend/babel-plugin-assets-api@1.4.0
+
+## 13.4.0
+
+### Minor Changes
+
+- 3bd16fc: feat: add preset modules option
+
+  if process.env.ESM is set to true then the babel config will set preset-env.modules option to `false` so the output will be esm.
+
+## 13.3.1
+
+### Patch Changes
+
+- 1abc22f: chore: upgrade dependencies
+- Updated dependencies [1abc22f]
+  - @talend/babel-plugin-import-from-index@1.7.1
+  - @talend/babel-plugin-assets-api@1.3.1
+
 ## 13.3.0
 
 ### Minor Changes