+++ tools/scripts-config-cdn/CHANGELOG.md
@@ -1,5 +1,28 @@
 # CHANGELOG
 
+## 10.12.0
+
+### Minor Changes
+
+- c3750a1: chore: upgrade dependencies
+
+  Major upgrade for all packages that have a peerDependency on react or react-dom. Those packages now ask for react@18 and react-dom@18.
+
+### Patch Changes
+
+- Updated dependencies [c3750a1]
+  - @talend/dynamic-cdn-webpack-plugin@14.0.0
+  - @talend/module-to-cdn@9.11.0
+
+## 10.11.1
+
+### Patch Changes
+
+- 1abc22f: chore: upgrade dependencies
+- Updated dependencies [1abc22f]
+  - @talend/dynamic-cdn-webpack-plugin@13.1.1
+  - @talend/module-to-cdn@9.10.1
+
 ## 10.11.0
 
 ### Minor Changes