+++ tools/scripts-utils/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @talend/scripts-utils
 
+## 2.3.0
+
+### Minor Changes
+
+- c3750a1: chore: upgrade dependencies
+
+  Major upgrade for all packages that have a peerDependency on react or react-dom. Those packages now ask for react@18 and react-dom@18.
+
+## 2.2.1
+
+### Patch Changes
+
+- 1abc22f: chore: upgrade dependencies
+
 ## 2.2.0
 
 ### Minor Changes